// libs/admin-shared/src/model/sysEmployeeResponse.ts
/**
 * Generated by orval v6.11.1 🍺
 * Do not edit manually.
 * 工业资源boss管理端【API接口文档】
 * 工业资源boss管理端
 */
import type { SysAccountRoleRelationResponse } from './sysAccountRoleRelationResponse';

/**
 * 员工表,返回对象模型
 */
export interface SysEmployeeResponse {
  /** 职员工号 */
  code?: string;
  /** 公司id */
  companyId?: string;
  /** 公司name */
  companyName?: string;
  /** 创建时间 */
  createTime?: string;
  /** id */
  id?: string;
  /** 职员名称 */
  name?: string;
  /** 创建人员 */
  operIdCreate?: string;
  /** 修改人员 */
  operIdMod?: string;
  /** 创建人员 姓名 */
  operNameCreate?: string;
  /** 修改人员姓名 */
  operNameMod?: string;
  /** 所属组织code */
  orgCode?: string;
  /** 所属组织id */
  orgId?: string;
  /** 所属组织Name */
  orgName?: string;
  /** 所属组织Code */
  orgType?: string;
  /** 手机号码 */
  phone?: string;
  /** 岗位id */
  postId?: string;
  /** 岗位名称 */
  postName?: string;
  /** 备注 */
  remark?: string;
  /** 角色 */
  roles?: SysAccountRoleRelationResponse[];
  /** 职员性别;{ m 男 w 女} */
  sex?: string;
  /** 职员性别;{ m 男 w 女} */
  sexDesc?: string;
  /** 状态;{normal:在职/正常,cancle:注销/删除,exit:离职}  */
  status?: string;
  /** 状态中文翻译 */
  statusDesc?: string;
  /** 员工所属 */
  type?: string;
  /** 修改时间 */
  updateTime?: string;
  /** 系统账号编号 */
  userCode?: string;
  /** 系统账号ID */
  userId?: string;
  /** 系统账号名称 */
  userName?: string;
}